requirements.txt
yfinance==0.2.28
pandas==2.1.4
google-cloud-bigquery==3.11.4
requests==2.31.0
pandas-gbq==0.23.1
